+++ src/lib/tempString.js
@@ -1,27 +1,27 @@
 import { logIn, signOut, googleLogIn, signUp } from './firebase.js';
 import { changeHash } from '../route.js';
 export const home = () => {
-    const bienvenido = `
+  const bienvenido = `
     <h1>Bienvenido</h1>
     <button class='btn-signout'>Cerrar Sesión</button>
     `;
-    const divElem = document.createElement('div');
-    divElem.setAttribute('id', 'bienvenido'); 
-    divElem.innerHTML = bienvenido;
-    divElem
-     .querySelector('.btn-signout')
-     .addEventListener('click', () => { 
-       signOut();     
-       divElem.style.display='none';     
-        document.getElementById('container').style.display='block';
-     });
- 
-     const contenido = document.getElementById('conteiner');
-     contenido.appendChild(divElem); 
-   }; 
+  const divElem = document.createElement('div');
+  divElem.setAttribute('id', 'bienvenido');
+  divElem.innerHTML = bienvenido;
+  divElem
+    .querySelector('.btn-signout')
+    .addEventListener('click', () => {
+      signOut();
+      divElem.style.display = 'none';
+      document.getElementById('container').style.display = 'block';
+    });
+
+  const contenido = document.getElementById('conteiner');
+  contenido.appendChild(divElem);
+};
 
 export const pageSignIn = () => {
-    const template = `<form class="register">
+  const template = `<form class="register">
                       <img src="imgs/logo.png" alt="pets">
                       <input class= "space" id="email2" type="email" placeholder="Ingresa e-mail" required>
                       <input class= "space" id="password2" type="password" placeholder="Ingresa contraseña" required>
@@ -30,50 +30,39 @@ export const pageSignIn = () => {
                       <button class="button google space" id= "google-button"><img src="imgs/google-logo.png" alt="Google">Iniciar sesión con Google</button>
                       <img src="imgs/pets.png" alt="pets">
                       <p>¿No tienes cuenta? <a id="registrate" href="#">Regístrate</a></p>  
-                      </form>`
-    const element = document.createElement('div');
-     element.innerHTML = template;
-     element 
-     .querySelector('#btn-log-in')
-     .addEventListener('click', event => {
-           event.preventDefault();
-          const email2 = element.querySelector("#email2").value;
-          const password2 = element.querySelector("#password2").value;
-            logIn(email2, password2); 
-         });
-    element 
+                    </form>`
+  const element = document.createElement('div');
+  element.innerHTML = template;
+
+  element
+    .querySelector('#btn-log-in')
+    .addEventListener('click', event => {
+      event.preventDefault();
+      const email2 = element.querySelector("#email2").value;
+      const password2 = element.querySelector("#password2").value;
+      logIn(email2, password2);
+      changeHash('/home')
+    });
+
+  element
     .querySelector('#google-button')
     .addEventListener('click', event => {
-        event.preventDefault();
-        googleLogIn();
-        });
-    
-    element
+      event.preventDefault();
+      googleLogIn();
+    });
+
+  element
     .querySelector('#registrate')
     .addEventListener('click', event => {
-        event.preventDefault();
-        changeHash('/signup')
-        });
+      event.preventDefault();
+      changeHash('/signup')
+    });
 
-    //  const container = document.getElementById('container');
-    //  container.appendChild(element); 
-     return element; 
-
-    // const btnSignIn = element.querySelector('#btn-log-in');
-    // btnSignIn.addEventListener('click', () => {
-    //   event.preventDefault();
-    //   btnLogIn();
-    // })
-    // const btnSignInGoogle = element.querySelector('#google-button');
-    // btnSignInGoogle.addEventListener('click', () => {
-    //   event.preventDefault();
-    //   btnGoogle();
-    // })
-    //return template;
-  };
+  return element;
+};
 
 export const pageSignUp = () => {
-    const template = `<form class="register">
+  const template = `<form class="register">
                       <img src="imgs/logo.png" alt="pets">
                       <input class= "space" id="usser" type="text" placeholder="Elige tu usuario" required>
                       <input class= "space" id="email" type="email" placeholder="Ingresa e-mail" required>
@@ -82,23 +71,44 @@ export const pageSignUp = () => {
                       <button class="button space" id= "btn-sign-up">Registrar</button>
                       <img src="imgs/pets.png" alt="pets">
                       </form>`
-    const element2 = document.createElement('div');
-    element2.innerHTML = template;
+  const element2 = document.createElement('div');
+  element2.innerHTML = template;
 
-    element2 
-     .querySelector('#btn-sign-up')
-     .addEventListener('click', event => {
-           event.preventDefault();
-          const email = document.querySelector("#email").value;
-          const password = document.querySelector("#password").value;
-            signUp(email, password);
-         });
+  element2
+    .querySelector('#btn-sign-up')
+    .addEventListener('click', event => {
+      event.preventDefault();
+      const email = document.querySelector("#email").value;
+      const password = document.querySelector("#password").value;
+      signUp(email, password);
+      changeHash('/signin')
+    });
 
-    // const btnSignUp = element2.querySelector('#btn-sign-up');
 
-    // btnSignUp.addEventListener('click', () => {
-    //   event.preventDefault();
-    //   btnRegister();
-    // });
-      return element2;
+  return element2;
+};
+
+
+///CREANDO TEMPLATE DE HOME
+export const pageHome = () => {
+  const template = `<head class="name-user">
+                      <p>Holi</p>
+                    </head>
+                    `
+  const element3 = document.createElement('div');
+  element3.innerHTML = template;
+
+  // element3
+  //   .querySelector('#btn-sign-up')
+  //   .addEventListener('click', event => {
+  //     event.preventDefault();
+  //     const email = document.querySelector("#email").value;
+  //     const password = document.querySelector("#password").value;
+  //     signUp(email, password);
+  //     changeHash('/signin')
+
+  //   });
+
+
+  return element3;
 };
